Move query-chunking into StudentModule and related ORM-objects
